Tests: put fixtures in their own file; test data from files.
